+++ sources/org/telegram/ui/Components/CircularProgressDrawable.java
@@ -63,7 +63,7 @@ public class CircularProgressDrawable extends Drawable {
     }
 
     private void updateSegment() {
-        getSegments((float) ((SystemClock.elapsedRealtime() - this.start) % 5400), this.segment);
+        getSegments((SystemClock.elapsedRealtime() - this.start) % 5400, this.segment);
     }
 
     public static void getSegments(float f, float[] fArr) {
